Add the runner file and the tests for the class game
